requirements.txt
openai
pandas
sqlalchemy
psycopg2-binary
pymysql
python-dotenv